refactor screenshot taking further: move webkit-specific code to engines.webkit package
